# doc/api/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
    AnnounceRequest:
      type: object
      required:
      - fileName
      - contentType
      - sizeBytes
      properties:
        fileName:
          type: string
          examples:
          - Rechnung.pdf
          pattern: \S
          maxLength: 255
        contentType:
          type: string
          examples:
          - application/pdf
          pattern: \S
          maxLength: 255
        sizeBytes:
          type: integer
          format: int64
          examples:
          - 48213
          minimum: 1
    AttachmentResponse:
      type: object
      required:
      - id
      - fileName
      - contentType
      - sizeBytes
      properties:
        id:
          type: integer
          format: int64
          examples:
          - 7
        fileName:
          type: string
          examples:
          - Rechnung.pdf
          maxLength: 255
        contentType:
          type: string
          examples:
          - application/pdf
          maxLength: 255
        sizeBytes:
          type: integer
          format: int64
          examples:
          - 48213
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    LinkResponse:
      type: object
      required:
      - url
      - expiresAt
      properties:
        url:
          type: string
          maxLength: 4096
        expiresAt:
          $ref: "#/components/schemas/Instant"
...
    RefusalResponse:
      type: object
      required:
      - refusal
      properties:
        refusal:
          type: string
          examples:
          - TASK_FULL
          enum:
          - UNSUPPORTED_TYPE
          - EMPTY
          - TOO_LARGE
          - TASK_FULL
          - USER_FULL
          - MISMATCH
...
    UploadResponse:
      type: object
      required:
      - attachment
      - url
      - headers
      - expiresAt
      properties:
        attachment:
          $ref: "#/components/schemas/AttachmentResponse"
        url:
          type: string
          maxLength: 4096
        headers:
          type: object
          additionalProperties:
            type: string
        expiresAt:
          $ref: "#/components/schemas/Instant"
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
      description: Authentication
...
paths:
...
  /api/tasks/{taskId}/attachments:
    post:
      summary: Announce an upload
      description: "Checks the type and the limits, and answers with a link that accepts\
        \ exactly this file: PUT it there with the given headers, then confirm."
      tags:
      - Attachments
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnounceRequest"
        required: true
      responses:
        "201":
          description: "The attachment, pending, and its upload link."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "422":
          description: "Refused: an unsupported type, too large, or a limit reached."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefusalResponse"
        "404":
          description: No task with this id is owned by the caller.
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme: []
  /api/tasks/{taskId}/attachments/{id}:
    delete:
      summary: Remove an attachment
      description: Deletes the file and its record.
      tags:
      - Attachments
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Removed.
        "404":
          description: No such attachment on a task the caller owns.
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tasks/{taskId}/attachments/{id}/confirm:
    post:
      summary: Confirm an upload
      description: Checks the file arrived as announced and makes the attachment available.
      tags:
      - Attachments
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: "The attachment, available."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttachmentResponse"
        "409":
          description: The file has not arrived yet.
        "422":
          description: What arrived is not what was announced; it was deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefusalResponse"
        "404":
          description: No such attachment on a task the caller owns.
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /api/tasks/{taskId}/attachments/{id}/link:
    get:
      summary: A link to open the file
      description: "Works for a few minutes, for the owner's browser to fetch the\
        \ file straight from storage."
      tags:
      - Attachments
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: The link and until when it works.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkResponse"
        "404":
          description: No such available attachment on a task the caller owns.
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
info:
  title: TaskFest API
  version: 1.0.0-SNAPSHOT
  description: The task board's REST surface. Every response schema here is published
    as JSON Schema and checked at runtime by the frontend.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
- url: /